requirements.txt
feedparser>=6.0.0
PyYAML>=6.0
python-dateutil>=2.8.0
scikit-learn>=1.0.0
numpy>=1.20.0
pytest>=7.0.0
